+++ config/app-console.php
@@ -3,7 +3,7 @@
 \Yii::setAlias('@app', dirname(__DIR__) . '/app');
 
 $db     = require __DIR__ . '/db.php';
-$routes  = require __DIR__ . '/routes.php';
+$params = require __DIR__ . '/params.php';
 
 $config = [
 	'id'                  => 'main-console',
@@ -54,17 +54,11 @@ $config = [
 		],
 		'db'    => $db,
 		'settings' => [
-			'class'     => 'justcoded\yii2\settings\components\DbSettings',
+			'class'     => 'app\components\Settings',
+			'defaults' => $params,
 			'modelsMap' => [
 				'app' => 'justcoded\yii2\settings\forms\AppSettingsForm',
 			],
-			'defaults' => [
-				'params' => [
-					'systemEmail' => ['admin@example.com' => 'Support'],
-					'adminEmail' => ['admin@example.com' => 'John Doe'],
-					'passwordResetTokenExpire' => 3600,
-				]
-			]
 		],
 	],
 ];